requirements.txt
confluent-kafka==2.13.2
Faker==40.12.0
psycopg2-binary==2.9.11
PyYAML==6.0.3